fix up profilers to use common parsed function data
